Use choice tables to make AST-generation decisions.
When replacing a TermHole or StmtHole, use a weighted-choice table to
decide on the replacement.  The table is obtained through an attribute
of the hole.

When the hole is at or below the maximum tree depth, we use a table that
allows for any possible production.  When the hole is above the maxmimum
tree depth, we use a table that forces the generator to choose an
"atomic" production: a Num for a Term, and an Eval for a Stmt.

This commit is just a refactoring: the choice tables are coded so that
the generator makes the same programs as it did before.