Support comments in the AST.
Every AST node now has a "precomment" child and a "postcomment" child.
The precomment is the body of a source-code comment (represented as a
`doc`) that is pretty-printed before the main content of the node.  The
postcomment is similar, except that it is pretty-printed after the main
content of the node.

The default pre- and postcomments are `empty`.  The empty document is
special-cased to print nothing (rather than "/* */", say).

`fresh-Prog` now creates a program with a precomment that identifies (1)
the version of the generator and (2) the value of the random seed.

The pretty-printing of comments "needs some work," let's say, but mostly
it is OK.