Rewrite `fresh-var-name` to avoid `gensym`.
Avoid calling `gensym` to generate fresh variable names.  The results of
`gensym` are random is a way that we do not control, and little changes
to our code cause changes in the symbols returned by `gensym`.

The new implementation of `fresh-var-name` is more stable with respect
to little changes in the code generator.