Encapsulate the mutable state of the generator.
Keep the mutable state of the program generator in a struct, and access
it via a parameter.  Now `do-it` is a function: if you call it twice
with the same options, it generates the same program twice.

At the moment, the state only includes the fresh-name counter.  It
should perhaps include the RNG and the options, too, but those things
are currently made available through other parameters.

Because I used a strucut, the current function names are a little
verbose (`set-generator-state-fresh-name-counter!`).  Feel free to
improve.

Anyway: there will be more state, and we will encapsulate it.