Add `make-do-it` for making many random programs.
`make-do-it` returns a closure that can be called multiple times, each
time outputting a new random program.

BUG: On the second and subsequent invocations, the seed printed in the
generated program is not the correct seed for recreating the program.
To fix this, we either need to (1) seed the RNG with a new seed on each
invocation or (2) output a "generation number" or similar along with the
original seed.