Improve the pretty-printing rules.
Greatly clean up the pretty-print attributes for most node types.
The result is much prettier output, but there is sill more to do!

Define function `v-comment` to "vertically wrap" a doc with the pre- and
postcomments of a node.  Similarly, define `h-comment` to "horizonally
wrap."  These functions are aware of cases in which one or both of the
comments are empty, and doesn't output empty documents in those cases
(which would turn into blank lines in the ultimate output).

Define more lexeme docs: comma, qmark, colon, keywords.
Postfix keyword doc variable name with ":", to avoid conflics with
Racket builtins (e.g., `if`).