rename `src` to `xsmith`, but symlink to `src`
This is to be easier to deal with packaging -- without extra work that
I would need to look up, Racket determines the name of a (locally
installed) package based on the directory name.  Since that is how I
will install it, I want the name to work out.

But I'm leaving a `src` symlink for convenience.