change scope-graph binding struct
It used to have a field that was a dictionary because I didn't know
what it needed.  This just makes those dictionary fields a part of the
struct since I know what they are now.  I suppose it could grow more
optional fields later?  But it will likely remain as-is.