Determine the git revision dynamically.
Determine the git commit hash dynamically, rather than (only) when the
`xsmith-version.rkt` file is compiled.  This is a more accurate way to
keep track of the curent git revision while working on Xsmith.  The
cost, of course, is that Xsmith tries to determine the git revision
every time it is run, not only when it is compiled.

I'm not sure that this is a good trade-off, but until I know that it is
a bad one, I'm going to opt for accuracy.