Rewrite the code to find the current git short hash.
Now we use three methods, in this order, to discern the short git hash
that corresponds to the running version of Xsmith:

  1. Look for the hash in `info.rkt`.  `git archive` will put in there,
     but most people won't run Xsmith from source trees produced in this
     way.

  2. Look for the hash in the package's checksum.  The Racket package
     catalog can provide us with the hash we want, and most users will
     install Xsmith through the package catalog.

  3. Look for the hash by running `git rev-parse` in the package's
     directory.  This works when the package's directory is part of a
     git repository's working tree.  Xsmith developers will commonly
     install Xsmith from a local directory that is part of a git
     repository's working tree.

If none of these methods work, the git commit hash is "unknown".

We no longer record the git commit hash at compile time.  This was a
pretty unreliable way of tracking the current git hash.