fix name access bug
The `fresh` property for definitions assumed the hole had fields for
name and type.  This is used in lifting -- the info has to be lifted
into a hole which gets fresh called on it later.  However, if the
definition node has a supertype node that is being filled, it may not
have those fields.

Thanks again to Pierce for writing code exercising a case I clearly
hadn't fully considered.