Make auto-wrapped distributions code work.
@pdarragh I left in all the commented code and some `eprintf` calls
that run at expansion time when using one of the wrapped identifiers.
In particular, knowing nothing about the library I don't know how they
deal with being called as identifier macros (IE as an identifier not
wrapped by parentheses, as is the case for the id `foo` in the form
`(map foo bar)`.  For things documented as procedures I imagine it
will work fine.  At any rate, you'll want to clean this up a bit, but
I figured leaving the other stuff will maybe make it easier to see the
core you were missing as a small, less cluttered diff.

Also I changed a `begin` to `let`.  The `begin` form splices into
definition contexts, and importantly does not create a definition
context.  Macros of the form:

```
    (define foo ...)
    ...
    <use-foo>)
```

fail when not used in a definition context.  So you want to use `let`
or something else that creates a definition context for those kinds of
macros.