Add option to print an s-expression representation on error.
Largely this is so I can have a standard representation of the tree to
debug rather than reading the rendered format.  Also with a little
more work it could be used to serialize and deserialize programs, but
that will probably never go anywhere, and with “parameterized
randomness” a la Zest we already have a feedback-driven mutation
strategy.