fix at-least-as-settled predicate used in type checking
`at-least-as-settled` is basically an optimization to say we can stop
exploring the tree because no further type unification is going to
change the answer we give to whether a node choice's type constraint
matches the hole.  It had a bug where it would consider a variable
with multiple base types sufficiently settled if the constraint type
was a single base type.  Now it checks further about those base types.