canned-components: fix typing for the “mutable” generic
I had written it as being covariant without really thinking about it.
But it needs to be invariant.  This allows (immutable (array X)) to be
covariant while forcing (mutable (array X)) to be invariant.
Famously, mutable arrays need to be invariant.  But I was allowing
mutable arrays to be covariant, and therefore generating bad code!