optimization: type-check the whole tree after each change
The old code tries to avoid having to do a whole-tree type check.
However, this ultimately leads to a lot of duplicated walking when
some features *do* need to walk a large part of the tree.  This commit
just type checks everything up front.  Due to RACR caching, it only
has to check the spine of the tree, which is much cheaper than the big
tree walks that need to be repeated with the other strategy.

I'm not really sure why I thought I had to avoid this earlier, because
it's way faster and much easier too...