Replace previous “optimization” with middle ground
The first optimization was sometimes great (eg. 5x better), but
sometimes worse (to a similar degree).  Many trials show that it was
close to a wash, with the “optimization” being a little worse on
average.  This patch tries to be more aggressive with type checking
that can be cached (like the optimization), but only when some of that
work is called for (like the old version).

I'm hopeful that this is a happy middle ground where we don't do the
work up-front but we also cache more so we aren't re-doing lots of
work in the worst case.  Quick tests show that for a given seed it can
be the fastest of the three, the slowest, or in the middle, but so far
it seems to avoid the worst-case performance of both other versions.

A longer benchmark shows that it is on average a 30% speed-up from the
original code.